+++ lerna-labeler/index.js
@@ -2,7 +2,7 @@ module.exports =
 /******/ (() => { // webpackBootstrap
 /******/ 	var __webpack_modules__ = ({
 
-/***/ 4627:
+/***/ 9391:
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
 const { info, warning } = __webpack_require__(2186);
@@ -45,14 +45,14 @@ module.exports.addLabelsToPr = addLabelsToPr;
 /***/ ((__unused_webpack_module, __unused_webpack_exports, __webpack_require__) => {
 
 const core = __webpack_require__(2186);
-const gitHub = __webpack_require__(5438);
+const github = __webpack_require__(5438);
 
 const { inputList } = __webpack_require__(2381);
 const { labeler } = __webpack_require__(7497);
 const { labelerSinceTag } = __webpack_require__(7637);
 
 const params = {
-  gitHubClient: gitHub.getOctokit(core.getInput('GITHUB_TOKEN')),
+  gitHubClient: github.getOctokit(core.getInput('GITHUB_TOKEN')),
   skipLabels: inputList(core.getInput('skip-labels')),
   dryRun: core.getInput('dry-run') === 'true',
   prefix: core.getInput('prefix'),
@@ -83,7 +83,7 @@ const { trueUpGitHistory } = __webpack_require__(8762);
 const { exec } = __webpack_require__(6264);
 const { cleanAppName, appNameEquals } = __webpack_require__(2381);
 
-const { addLabelsToPr } = __webpack_require__(4627);
+const { addLabelsToPr } = __webpack_require__(9391);
 
 async function changedSinceLatestTag(rootDir, packageName, path) {
   const tagPrefix = `refs/tags/${packageName}@*`;
@@ -104,10 +104,10 @@ async function changedSinceLatestTag(rootDir, packageName, path) {
 
 async function findChangedPackages(project) {
   const pkgJsons = await project.getPackages();
-  const rootDir = project.packageParentDirs[0].split('/').pop();
+  const projectDir = project.packageParentDirs[0].split('/').pop();
 
   const packages = await Promise.all(
-    pkgJsons.map((pkgJson) => changedSinceLatestTag(rootDir, pkgJson.name, pkgJson.location.split('/').pop())),
+    pkgJsons.map((pkgJson) => changedSinceLatestTag(projectDir, pkgJson.name, pkgJson.location.split('/').pop())),
   );
 
   return packages.filter(Boolean);
@@ -141,35 +141,44 @@ module.exports.labelerSinceTag = labelerSinceTag;
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
 const { info } = __webpack_require__(2186);
+const github = __webpack_require__(5438);
 const Project = __webpack_require__(234);
 
 const { LABEL_PREFIX } = __webpack_require__(1404);
-const { getDestBranch, getSrcBranch } = __webpack_require__(8762);
+const { getDestBranch, getSrcBranch, trueUpGitHistory } = __webpack_require__(8762);
 const { exec } = __webpack_require__(6264);
 const { cleanAppName, appNameEquals } = __webpack_require__(2381);
 
-const { addLabelsToPr } = __webpack_require__(4627);
+const { addLabelsToPr } = __webpack_require__(9391);
 
 async function findChangedFiles(srcBranch, destBranch) {
-  const changes = await exec(`git diff --name-only ${destBranch} ${srcBranch}`);
+  if (github.context.actor) {
+    // Needed for finding the current branch
+    await trueUpGitHistory();
+  }
+
+  const remote = await exec('git remote');
+  const changes = await exec(`git diff --name-only ${remote ? `${remote}/` : ''}${destBranch} ${srcBranch}`);
 
   return changes.split('\n').filter(Boolean);
 }
 
 async function findChangedPackages(project, changedFiles) {
   const pkgJsons = await project.getPackages();
-  const rootDir = project.packageParentDirs[0].split('/').pop();
+  const projectDir = project.packageParentDirs[0].split('/').pop();
 
   const changedPkgDirs = changedFiles
-    .filter((filename) => filename.startsWith(rootDir))
-    .map((filename) => filename.substring(0, filename.indexOf('/', rootDir.length + 1)));
+    .filter((filename) => filename.startsWith(projectDir))
+    .map((filename) => filename.substring(0, filename.indexOf('/', projectDir.length + 1)));
 
   const uniqChangedPkgDirs = [...new Set(changedPkgDirs)];
 
-  return uniqChangedPkgDirs.map((pkgDir) => {
-    const { name } = pkgJsons.find((pkgJson) => pkgJson.location.endsWith(pkgDir));
-    return name;
-  });
+  return uniqChangedPkgDirs
+    .map((pkgDir) => {
+      const { name } = pkgJsons.find((pkgJson) => pkgJson.location.endsWith(pkgDir)) || {};
+      return name;
+    })
+    .filter(Boolean);
 }
 
 async function labeler({ gitHubClient, skipLabels = [], dryRun = false, prefix = LABEL_PREFIX }) {
@@ -191,9 +200,11 @@ async function labeler({ gitHubClient, skipLabels = [], dryRun = false, prefix =
   info(changedFiles.join('\n'));
 
   const packages = await findChangedPackages(project, changedFiles);
+
   const labels = packages
     .filter((name) => !skipLabels.some((label) => appNameEquals(label, name)))
-    .map((name) => `${prefix}${cleanAppName(name, prefix)}`); // remove @scope/ and add deploy: prefix
+    .map((name) => `${prefix}${cleanAppName(name, prefix)}`) // remove @scope/ and add deploy: prefix
+    .sort();
 
   if (!dryRun) {
     await addLabelsToPr(gitHubClient, labels);
@@ -54404,6 +54415,8 @@ module.exports.LABEL_PREFIX = 'deploy:';
 /***/ 8762:
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
+/* eslint-disable camelcase */
+/* eslint-disable no-await-in-loop */
 const { info } = __webpack_require__(2186);
 const github = __webpack_require__(5438);
 
@@ -54413,32 +54426,26 @@ const ENV_BRANCHES = ['master', 'qa', 'prod', 'hc'];
 
 async function getShortCommit() {
   if (github.context.payload) {
-    /* eslint-disable camelcase */
     const { pull_request } = github.context.payload;
     const sha = !pull_request || pull_request.merged ? github.context.sha : pull_request.head.sha;
     return sha.substring(0, 8);
-    /* eslint-enable camelcase */
   }
 
   return exec('git rev-parse --short=8 HEAD');
 }
 
 async function getDestBranch() {
-  /* eslint-disable camelcase */
   const { pull_request } = github.context.payload;
   const branch = (pull_request && pull_request.base && pull_request.base.ref) || github.context.ref;
-  /* eslint-enable camelcase */
 
   return branch ? branch.split('/').pop() : exec('git rev-parse --abbrev-ref HEAD');
 }
 
 async function getSrcBranch() {
-  /* eslint-disable camelcase */
   const { pull_request } = github.context.payload;
   if (pull_request) {
     return pull_request.head.ref;
   }
-  /* eslint-enable camelcase */
 
   const branch = github.context.ref
     ? github.context.ref.split('/').pop()
@@ -54516,10 +54523,8 @@ async function trueUpGitHistory() {
   if (isShallowFetch) {
     const destBranch = await getDestBranch();
     const srcBranch = await getSrcBranch();
-    /* eslint-disable camelcase */
     const { pull_request } = github.context.payload;
     const merged = pull_request && pull_request.merged;
-    /* eslint-enable camelcase */
 
     await sh(
       `git fetch --prune --unshallow --tags
@@ -54555,12 +54560,9 @@ async function gitMerge(params = {}) {
 
   // eslint-disable-next-line no-restricted-syntax
   for (const destBranch of destBranches) {
-    // eslint-disable-next-line no-await-in-loop
-    await sh(
-      `git checkout ${destBranch}
-      git merge ${srcBranch}
-      git push "${gitUrl}" --follow-tags`,
-    );
+    await exec(`git checkout ${destBranch}`, { echo: true });
+    await exec(`git merge ${srcBranch}`, { echo: true });
+    await exec(`git push "${gitUrl}" --follow-tags`, { echo: true });
 
     srcBranch = destBranch;
   }
@@ -54628,7 +54630,10 @@ async function sh(cmd) {
   });
 }
 
-async function exec(cmd) {
+async function exec(cmd, { echo } = { echo: false }) {
+  if (echo) {
+    info(cmd);
+  }
   const { stdout } = await execPromise(cmd);
 
   return stdout.trim();