+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.4](https://github.com/risk-and-safety/platform/compare/github-actions@3.2.3...github-actions@3.2.4) (2021-01-12)
+
+
+### Bug Fixes
+
+* **deploy:** catch npm publish errors, allow npm and GitHub ([18e9102](https://github.com/risk-and-safety/platform/commit/18e910278eb07e49e166fe3a460a46dcf23c9200))
+
+
+
+
+
+## [3.2.3](https://github.com/risk-and-safety/platform/compare/github-actions@3.2.2...github-actions@3.2.3) (2021-01-11)
+
+**Note:** Version bump only for package github-actions
+
+
+
+
+
 ## [3.2.2](https://github.com/risk-and-safety/platform/compare/github-actions@3.2.1...github-actions@3.2.2) (2020-12-18)